babzel-tools/src/test/java/org/babzel/tools/opennlp/unidep/util/UniDepConlluModelFileSupplierTest.java
/*
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package org.babzel.tools.opennlp.unidep.util;

import com.google.common.jimfs.Jimfs;
import static org.assertj.core.api.Assertions.assertThat;
import org.babzel.tools.util.ResourceDirectorySupplier;
import org.babzel.tools.util.RootDirectorySupplier;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.BDDMockito.given;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;

@ExtendWith(MockitoExtension.class)
public class UniDepConlluModelFileSupplierTest {
    @Mock
    private RootDirectorySupplier rootDirectorySupplier;
    @Mock
    private ResourceDirectorySupplier resourceDirectorySupplier;
    @InjectMocks
    private UniDepConlluModelFileSupplier supplier;

    @Test
    public void getFile() {
        var rootPath = Jimfs.newFileSystem().getPath("");
        given(rootDirectorySupplier.getRootDirectory()).willReturn(rootPath);
        var resPath = rootPath.resolve("res");
        given(resourceDirectorySupplier.getResourceDirectory(any())).willReturn(resPath);

        assertThat(supplier.getTreebankFile()).isEqualTo(rootPath.resolve("uni-dep-treebank.tgz"));
        assertThat(supplier.getConlluFile("lx")).isEqualTo(rootPath.resolve("lx.conllu"));
        assertThat(supplier.getSentenceModelFile("lx")).isEqualTo(resPath.resolve("lx-sentence-detector.onlpm"));
        assertThat(supplier.getSentenceReportFile("lx")).isEqualTo(resPath.resolve("lx-sentence-detector.txt"));
        assertThat(supplier.getTokenizerModelFile("lx")).isEqualTo(resPath.resolve("lx-tokenizer.onlpm"));
        assertThat(supplier.getTokenizerReportFile("lx")).isEqualTo(resPath.resolve("lx-tokenizer.txt"));
        assertThat(supplier.getTokenExpanderModelFile("lx")).isEqualTo(resPath.resolve("lx-token-expander.onlpm"));
        assertThat(supplier.getTokenExpanderReportFile("lx")).isEqualTo(resPath.resolve("lx-token-expander.txt"));
        assertThat(supplier.getPOSModelFile("lx")).isEqualTo(resPath.resolve("lx-pos-tagger.onlpm"));
        assertThat(supplier.getPOSReportFile("lx")).isEqualTo(resPath.resolve("lx-pos-tagger.txt"));
        assertThat(supplier.getLemmatizerModelFile("lx")).isEqualTo(resPath.resolve("lx-lemmatizer.onlpm"));
        assertThat(supplier.getLemmatizerReportFile("lx")).isEqualTo(resPath.resolve("lx-lemmatizer.txt"));
    }
}
